Add files via upload
Unfortunately, despite advancements in research and treatment, nowadays, Breast cancer is one of the most diagnosed diseases worldwide, with more than 1.6 million new cases each year among women.
Now, before diving into the serious stuff, I ask you: “Try to imagine you walk into a doctor’s office, and he straight away tells you you got breast cancer”.
Now, if you’re a guy, just like me, for saying, you may be trying to call yourself out of this “particular” situation. Probably, you’re feeling like you’re missing something. Well, unfortunately, boys, we must clear up a common misconception: Yes, men can get breast cancer, too! Shocking, right?
But never mind, because today we’re focusing on female patients thanks to data from the SEER Program. As you can see here, we have 4024 observations and 16 features!

A crucial first step in our analysis was identifying outliers. As you can see in the boxplot visualizations, there are so many! Using the Interquartile Range (IQR) Method, which provides a clear way to detect anomalies using quantiles (I’m sure everyone knows it), we found 617 outliers (that’s 15% of the dataset). Since they represent a great portion of information, we decided to keep them in order to preserve data integrity.
We examined various clinical features, such as age, race, and marital status. Some of them are even more specific to the domain, like: T Stage, N Stag,e and more. However, none of them are normally distributed, as you can see in the bar plots.
We also checked for Multicollinearity, and it seems like there are no strong correlations between numerical features.

Imbalanced target class
Here, we present the main challenge in our analysis—and a common issue in many clinical datasets: class imbalance. As you can see, only 15% of the patients fall into the 'dead' category, which is the class we aim to predict. It’s clear that such an imbalance could heavily impact the performance of our machine learning models if not addressed properly. We will explain how we handled this issue later in the presentation. For now, let’s move on and introduce the Machine Learning section.
Baseline ML Models
We tested two models: Logistic Regression and Random Forest, to provide a fair comparison. The choice of Logistic Regression was almost mandatory, as it is well-suited for predicting binary target variables, like in our case. Moreover, it is one of the first generalized linear models we are typically introduced to in statistical learning.
Here, we present the main challenge in our analysis—and a common issue in many clinical datasets: class imbalance. As you can see, only 15% of the patients fall into the 'dead' category, which is the class we aim to predict. It’s clear that such an imbalance could heavily impact the performance of our machine learning models if not addressed properly. We will explain how we handled this issue later in the presentation. For now, let’s move on and introduce the Machine Learning section.
Baseline ML Models
We tested two models: Logistic Regression and Random Forest, to provide a fair comparison. The choice of Logistic Regression was almost mandatory, as it is well-suited for predicting binary target variables, like in our case. Moreover, it is one of the first generalized linear models we are typically introduced to in statistical learning.

After oversampling, Logistic Regression improved recall for the "Deceased" class, sacrificing overall accuracy, while Random Forest maintained overall accuracy.
However, neither model achieved optimal results, but I think statistical thinking matters the most in this case.
So, this is a brief recap of our analysis. We can clearly say that Logistic Regression was the most effective model for our specific task.
If we wanted to enhance the model’s performance, we could add more relevant features to the dataset, collect more observations for the minority class, or explore models that better handle imbalanced data.